logical assignment plugin, fix inline conditions
