+++ designer/babel.config.js
@@ -27,6 +27,8 @@ module.exports = {
     "@babel/plugin-proposal-export-default-from",
     "@babel/plugin-proposal-nullish-coalescing-operator",
     "@babel/plugin-proposal-optional-chaining",
+    "@babel/plugin-proposal-logical-assignment-operators",
+
     [
       "module-resolver",
       {
