+++ package.json
@@ -40,6 +40,7 @@
     "@babel/eslint-plugin": "^7.11.3",
     "@babel/plugin-proposal-class-properties": "7.10.1",
     "@babel/plugin-proposal-export-default-from": "^7.10.1",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.13.8",
     "@babel/plugin-proposal-private-methods": "^7.10.1",
     "@babel/plugin-proposal-private-property-in-object": "^7.10.1",
     "@babel/plugin-transform-runtime": "^7.11.0",
